Add instance-level UUID to health check logs
Generate a UUID when the health check is initiated
and include it in the health check logs. This UUID
is used to query Elasticsearch for the timestamp
of the last available health check log. The addition
of the UUID allows multiple pods to use the same
configuration while storing logs in the same
Elasticsearch database,